Refactor install script
It now properly creates a backup of existing files before overriding
them.

Assuming you are using a POSIX compliant operating system, only git is
required to install (and use/configure!) my dotfiles.

If rsync is available, it will use it as it is still significantly
faster than the alternative method to moving the files to their correct
location.