Improve jq formatting selection
If for some reason you modified the formatting of the json blob, jq
will error out even if the whole thing is still technically valid
json. This is because the mark selection process would only try to
format that specific lines changes, which is not ideal for a json
file with only 1 json blob. Forcing it to select the entire json
blob is less error-prone.